corrections

+++ src/main/java/com/example/FileCsv/Repo/FileRepo.java
@@ -3,5 +3,6 @@ package com.example.FileCsv.Repo;
 import com.example.FileCsv.Entity.Model;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FileRepo extends JpaRepository<Model Long> {
+public interface FileRepo extends JpaRepository<Model ,Integer> {
+
 }
